+++ src/dipdup/package.py
@@ -4,14 +4,14 @@ from collections.abc import Awaitable
 from collections.abc import Callable
 from collections.abc import Generator
 from pathlib import Path
-from typing import TYPE_CHECKING
 from typing import Any
-from typing import TypedDict
 from typing import cast
 
 from pydantic import BaseModel
 
 from dipdup import env
+from dipdup.abi.cairo import CairoAbiManager
+from dipdup.abi.evm import EvmAbiManager
 from dipdup.exceptions import ProjectPackageError
 from dipdup.project import Answers
 from dipdup.project import answers_from_replay
@@ -19,11 +19,6 @@ from dipdup.utils import import_from
 from dipdup.utils import import_submodules
 from dipdup.utils import pascal_to_snake
 from dipdup.utils import touch
-
-if TYPE_CHECKING:
-    from starknet_py.cairo.data_types import CairoType  # type: ignore[import-untyped]
-    from starknet_py.serialization import PayloadSerializer  # type: ignore[import-untyped]
-
 
 KEEP_MARKER = '.keep'
 PACKAGE_MARKER = '__init__.py'
@@ -57,36 +52,6 @@ def draw_package_tree(root: Path, project_tree: dict[str, tuple[Path, ...]]) -> 
     return tuple(lines)
 
 
-class ConvertedEventAbi(TypedDict):
-    name: str
-    topic0: str
-    inputs: tuple[tuple[str, bool], ...]
-    topic_count: int
-
-
-class ConvertedMethodAbi(TypedDict):
-    name: str
-    sighash: str
-    inputs: tuple[dict[str, str], ...]
-    outputs: tuple[dict[str, str], ...]
-
-
-class ConvertedEvmAbi(TypedDict):
-    events: dict[str, ConvertedEventAbi]
-    methods: dict[str, ConvertedMethodAbi]
-
-
-class ConvertedEventCairoAbi(TypedDict):
-    name: str
-    event_identifier: str
-    members: dict[str, 'CairoType']
-    serializer: 'PayloadSerializer'
-
-
-class ConvertedCairoAbi(TypedDict):
-    events: dict[str, ConvertedEventCairoAbi]
-
-
 class DipDupPackage:
     def __init__(self, root: Path) -> None:
         _logger.info('Loading package `%s` from `%s`', root.name, root)
@@ -114,8 +79,8 @@ class DipDupPackage:
         self._replay: Answers | None = None
         self._callbacks: dict[str, Callable[..., Awaitable[Any]]] = {}
         self._types: dict[str, type[BaseModel]] = {}
-        self._converted_evm_abis: dict[str, ConvertedEvmAbi] = {}
-        self._converted_cairo_abis: dict[str, ConvertedCairoAbi] = {}
+        self._evm_abis = EvmAbiManager(self)
+        self._cairo_abis = CairoAbiManager(self)
 
     @property
     def cairo_abi_paths(self) -> Generator[Any, None, None]:
@@ -176,6 +141,10 @@ class DipDupPackage:
 
         self._post_init()
 
+    def load_abis(self) -> None:
+        self._evm_abis.load()
+        self._cairo_abis.load()
+
     def _pre_init(self) -> None:
         if self.name != pascal_to_snake(self.name):
             raise ProjectPackageError(f'`{self.name}` is not a valid Python package name')
@@ -218,17 +187,3 @@ class DipDupPackage:
                 raise ProjectPackageError(f'`{path}.{name}` is not a valid callback')
             self._callbacks[key] = callback
         return cast(Callable[..., Awaitable[None]], callback)
-
-    def get_converted_evm_abi(self, typename: str) -> ConvertedEvmAbi:
-        if not self._converted_evm_abis:
-            from dipdup.codegen.evm import convert_abi
-
-            self._converted_evm_abis = convert_abi(self)
-        return self._converted_evm_abis[typename]
-
-    def get_converted_starknet_abi(self, typename: str) -> ConvertedCairoAbi:
-        if not self._converted_cairo_abis:
-            from dipdup.codegen.starknet import convert_abi
-
-            self._converted_cairo_abis = convert_abi(self)
-        return self._converted_cairo_abis[typename]